state即使是整型也要保存为小数类型，才能够保证能够进行计算

+++ src/envs/gym.py
@@ -62,7 +62,7 @@ class Gym(SingleAgentEnv):
             state = np.array(state)
         if len(state.shape) == 0:
             state = np.expand_dims(state, axis=0)
-        state = torch.tensor(state)
+        state = torch.tensor(state).float()
         self.state = state
 
     def get_state(self):
